Revert "use updated options dict".
Context: The options dict is supposed to be global and provided by the user, in the form of a dict mapping block names to dicts of options for those blocks, and is not updated by kwargs; instead, it is combined with them for the current block call and also passed on to calls to the same block like .steady_state() after .solve_steady_state(), etc. This commit inadvertently passed on a flattened options dict that is no longer usable by child blocks.

This reverts commit 87942491a57df99a7d3158810a8ec42d3f2034f9.